Set up database init script and interface service
The interface service uses an express app to listen for http posts, which are called by the functions the module (same file) exports. So, we're relying on additional ports and http communication, but no other module we write needs to worry about those details.

Note that the API that db.service exports does NOT accommodate asynchronous behavior. If a script wishes to perform successive operations on the same document which must be performed in a certain order, it'll probably have to connect to the database itself.

Resolves #3