Revise DB interface module to not busy communication lines
Brilliant as my module/service solution was, it was rather unwieldy, and today I realized it didn't allow for very convenient access to MongoDB cursors, which are kind of important. The new technique is directly off of someone on Stack Overflow who sounded like they knew what they were talking about. This one requires the client service to do everything inside a connection callback, and I'm not sure yet if it works as desired when used by multiple independent services. But it's unquestionably far more elegant.

Contributes to #3