Fill Standard LO collection with real data
The 'accreditation' directory is to contain raw and processed documentatioin for Standard LOs. The git repository itself need only have the tabularized .dat files from which a node script (db-NCWF.js, for this commit) inserts records into the database.

The perl script in this commit is responsible for getting from the raw documentation (not included) to the (included) tabularized .dat file. If we need to repeat the process, detailed instructions are included in the script itself.

lo-suggestion's change is just in its testing space.

Contributes to #2