Finish script to parse mysql heroku data and add to mongo db
Now fully TypeScripted, and ostensibly finished. But note it ISN'T tested. Before it will actually work, there are certainly bugs that will need to be sorted out...

A couple of the changes in package.json are related to my next commit, which is updating the other scripts to TypeScript.

Contributes to #1, #4