:fire: Remove unneeded code
The catch statement in the createChangeLog function was catching anything thrown and converting it to an error if it wasn't already.
This is no longer needed, as all function calls in the try block will throw a ResourceError or ServiceError. Therefore, was can just allow this
to fall through to the route handler.